Allow only 8 dual-use layer keys
Instead of 24, only allow 8 layer keys. This is in line with recent
changes to OneShot, and the goal is to reduce code size, considerably.

In our case, we save almost 150 bytes just with this tiny change, and 8
dual-use layer keys are still plenty.

+++ src/Kaleidoscope/DualUse.cpp
@@ -22,8 +22,8 @@
 using namespace KaleidoscopePlugins::Ranges;
 
 namespace KaleidoscopePlugins {
-  uint32_t DualUse::keyActionNeededMap;
-  uint32_t DualUse::pressedMap;
+  uint16_t DualUse::keyActionNeededMap;
+  uint16_t DualUse::pressedMap;
   uint32_t DualUse::endTime;
   uint16_t DualUse::timeOut = 1000;
